update backend
full code backend

+++ controllers/blogController.js
@@ -1,7 +1,7 @@
 import Blog from '../models/blogModel.js';
 
 // Get all blogs
-export const getBlogs = async (req, res) => {
+const getBlogs = async (req, res) => {
   try {
     const blogs = await Blog.find().populate('categories').populate('images');
     res.status(200).json(blogs);
@@ -11,7 +11,7 @@ export const getBlogs = async (req, res) => {
 };
 
 // Get a single blog by ID
-export const getBlogById = async (req, res) => {
+const getBlogById = async (req, res) => {
   try {
     const blog = await Blog.findById(req.params.id).populate('categories').populate('images');
     if (!blog) {
@@ -24,10 +24,16 @@ export const getBlogById = async (req, res) => {
 };
 
 // Create a new blog
-export const createBlog = async (req, res) => {
+const createBlog = async (req, res) => {
+  const isAdmin = req.user.role === 'admin';
+  const userId = req.user._id;
   const { title, content, categories, images, status } = req.body;
+  if (!isAdmin) {
+    return res.status(403).json({ message: 'You are not authorized to create a blog' });
+  }
   try {
     const newBlog = new Blog({
+      user: userId,
       title,
       content,
       categories,
@@ -42,8 +48,13 @@ export const createBlog = async (req, res) => {
 };
 
 // Update an existing blog
-export const updateBlog = async (req, res) => {
+const updateBlog = async (req, res) => {
   const { title, content, categories, images, status } = req.body;
+  const userId = req.user._id;
+  const isAdmin = req.user.role === 'admin';
+  if ( userId !== req.blog.user.toString() || !isAdmin ) {
+    return res.status(403).json({ message: 'You are not authorized to update this blog' });
+  }
   try {
     const updatedBlog = await Blog.findByIdAndUpdate(
       req.params.id,
@@ -60,7 +71,12 @@ export const updateBlog = async (req, res) => {
 };
 
 // Delete a blog
-export const deleteBlog = async (req, res) => {
+const deleteBlog = async (req, res) => {
+  const userId = req.user._id;
+  const isAdmin = req.user.role === 'admin';
+  if ( userId !== req.blog.user.toString() || !isAdmin ) {
+    return res.status(403).json({ message: 'You are not authorized to delete this blog' });
+  }
   try {
     const deletedBlog = await Blog.findByIdAndDelete(req.params.id);
     if (!deletedBlog) {
@@ -71,3 +87,5 @@ export const deleteBlog = async (req, res) => {
     res.status(500).json({ message: 'Error deleting blog', error });
   }
 };
+
+export { getBlogs, getBlogById, createBlog, updateBlog, deleteBlog };
